[ALPS08678993] mml:build error fix
Fix 32bit sys build error.

MTK-Commit-Id: bc77bfb59c98f77e05a55631e3a1c751b2629252

CR-Id: ALPS08678993
Feature: mml

+++ drivers/misc/mediatek/cmdq/mailbox/cmdq-test.c
@@ -1179,10 +1179,10 @@ static void cmdq_test_mbox_prebuilt_instr_ext_table(struct cmdq_test *test,
 	}
 
 	cmdq_msg("%s: pkt:%p pa:%#lx cmd_buf_size:%#lx pc:%#lx end:%#lx",
-		__func__, pkt, (unsigned long)buf->pa_base, pkt->cmd_buf_size,
+		__func__, pkt, (unsigned long)buf->pa_base, (unsigned long)pkt->cmd_buf_size,
 		(unsigned long)CMDQ_REG_SHIFT_ADDR(buf->pa_base),
 		(unsigned long)CMDQ_REG_SHIFT_ADDR(buf->pa_base +
-			pkt->cmd_buf_size));
+			(unsigned long)pkt->cmd_buf_size));
 
 	for (i = 0; i < pkt->cmd_buf_size / CMDQ_INST_SIZE; i++)
 		cmdq_msg(",%d,%#llx,", i, *((u64 *)buf->va_base + i));
@@ -1321,10 +1321,10 @@ static void cmdq_test_mbox_prebuilt_instr(struct cmdq_test *test,
 	}
 
 	cmdq_msg("%s: pkt:%p pa:%#lx cmd_buf_size:%#lx pc:%#lx end:%#lx",
-		__func__, pkt, (unsigned long)buf->pa_base, pkt->cmd_buf_size,
+		__func__, pkt, (unsigned long)buf->pa_base, (unsigned long)pkt->cmd_buf_size,
 		(unsigned long)CMDQ_REG_SHIFT_ADDR(buf->pa_base),
 		(unsigned long)CMDQ_REG_SHIFT_ADDR(buf->pa_base +
-			pkt->cmd_buf_size));
+			(unsigned long)pkt->cmd_buf_size));
 
 	for (i = 0; i < pkt->cmd_buf_size / CMDQ_INST_SIZE; i++)
 		cmdq_msg(",%d,%#llx,", i, *((u64 *)buf->va_base + i));
